Android gcc compilant types

+++ rand.cpp
@@ -14,7 +14,7 @@ srand(this->genSeed());
 #endif
 }
 
-uint32_t Rand::getTicks()
+unsigned long Rand::getTicks()
 {
 // TODO: Find a cpu time/ticks asm inline for ARMv6 and ARMv7
 #ifdef __arm__
@@ -23,7 +23,7 @@ uint32_t Rand::getTicks()
 
 	return usecs.tv_usec;
 #else
-	uint32_t tsc;
+	unsigned long tsc;
 	__asm__ __volatile__(
 			"rdtscp;"
 			"shl $32, %%rdx;"
